Internationalization and Trie optimization
* Change Trie to be Character based
* Remove US/UK distinction from Trie
* Combine the word indicator and number of children into one byte
* Use a single UTF-8 String to represent child nodes
* Eliminate case conversion - everything is in "how it looks onscreen" case
* Add dictionaries for OTCWL2016, SOWPODS, and Spanish

Internationalization is based on a Letter Chain generator. Rather than
worrying about digraphs in each language (they are impossible to discern
from a simple word list), digraphs are explicitly not supported. Instead,
a board is generated by creating probabilistic words in a transitionable
pattern. Probabilistic words are generated using the probability that
one letter follows another in a given dictionary, along with the word
length probabilities. This ensures, for example, that a "Q" always has
a "U" next to it if all words in the dictionary have that pattern,
unless the Q happens to end up as the last letter in a probabilistic
word.

The DictionaryParser class takes word lists and creates a compressed
Trie and the statistics used by the Letter Chain generator, along with
a point value for each letter based on letter frequencies.